requirements.txt
pandas~=2.2
requests~=2.32
beautifulsoup4~=4.12
SQLAlchemy~=2.0
psycopg2-binary~=2.9
gspread~=5.12  # Versi gspread yang lebih baru
google-auth~=2.28 # Versi google-auth yang lebih baru
google-api-python-client~=2.126 # Versi client API Google yang lebih baru
python-dotenv~=1.0
# For testing
pytest~=7.4
pytest-cov~=5.0
requests-mock~=1.11